moved login into journalist_app/main.py
