+++ install_files/ansible-base/roles/build-securedrop-app-code-deb-pkg/files/usr.sbin.apache2
@@ -135,6 +135,8 @@
   /var/www/securedrop/journalist_app/decorators.pyc rw,
   /var/www/securedrop/journalist_app/forms.py r,
   /var/www/securedrop/journalist_app/forms.pyc rw,
+  /var/www/securedrop/journalist_app/main.py r,
+  /var/www/securedrop/journalist_app/main.pyc rw,
   /var/www/securedrop/journalist_app/utils.py r,
   /var/www/securedrop/journalist_app/utils.pyc rw,
   /var/www/securedrop/journalist_app/__pycache__/** rw,
